requirements.txt
streamlit==1.51.0
tensorflow-cpu
numpy>=1.26.0,<2.0.0
pandas==2.3.3
pillow==12.0.0
opencv-python-headless
requests==2.32.5
langchain-google-genai
google-generativeai
SpeechRecognition
gTTS
deep-translator

python-dotenv
streamlit-option-menu
